PLAT-48441: Update enact-samples to react 16 (#45)
* Update samples to react 16 and some cleanup

* Update enact-all-samples and related paths

* clean up package.json files

* Restore entry points for enact-all-samples

Enyo-DCO-1.1-Signed-off-by: Stephen Choi <stephen.choi@lge.com>

* Remove deprecated props and fix lint warning

* update enact package to "next"
