+++ enact-all-samples/package.json
@@ -15,7 +15,7 @@
     "test-json": "enact test start --single-run --browsers PhantomJS --reporters json",
     "test-watch": "enact test start --browsers PhantomJS"
   },
-  "license": "UNLICENSED",
+  "license": "Apache-2.0",
   "private": true,
   "repository": "",
   "enact": {
@@ -28,15 +28,15 @@
     "extends": "enact"
   },
   "dependencies": {
-    "@enact/core": "^1.1.0",
-    "@enact/i18n": "^1.1.0",
-    "@enact/moonstone": "^1.1.0",
-    "@enact/spotlight": "^1.1.0",
-    "@enact/ui": "^1.1.0",
-    "@enact/webos": "^1.1.0",
-    "prop-types": "^15.5.8",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
+    "@enact/core": "next",
+    "@enact/i18n": "next",
+    "@enact/moonstone": "next",
+    "@enact/spotlight": "next",
+    "@enact/ui": "next",
+    "@enact/webos": "next",
+    "prop-types": "^15.6.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
     "react-redux": "^5.0.3",
     "react-router": "^4.1.1",
     "react-router-dom": "^4.1.1",
