+++ pattern-activity-panels-deep-linking/package.json
@@ -29,15 +29,15 @@
     "extends": "enact/strict"
   },
   "dependencies": {
-    "@enact/core": "^1.0.0",
-    "@enact/i18n": "^1.0.0",
-    "@enact/moonstone": "^1.0.0",
-    "@enact/spotlight": "^1.0.0",
-    "@enact/ui": "^1.0.0",
-    "@enact/webos": "^1.0.0",
-    "prop-types": "^15.5.8",
-    "react": "^15.5.0",
-    "react-dom": "^15.5.0",
+    "@enact/core": "next",
+    "@enact/i18n": "next",
+    "@enact/moonstone": "next",
+    "@enact/spotlight": "next",
+    "@enact/ui": "next",
+    "@enact/webos": "next",
+    "prop-types": "^15.6.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
     "react-redux": "^5.0.3",
     "redux": "^3.6.0"
   }
