+++ pattern-activity-panels-deep-linking/src/index.js
@@ -1,24 +1,16 @@
 import React from 'react';
 import {render} from 'react-dom';
-import {Provider} from 'react-redux';
-import App from './App';
-import configureStore from './store';
 
-// set default launch path
-const launchParam = '/first/second';
-const store = configureStore({
-	path: launchParam
-});
+import App from './main';
 
-const appElement = (
-	<Provider store={store}>
-		<App />
-	</Provider>
-);
+let appElement = <App />;
 
-// In a browser environment, render the app to the document.
+// In a browser environment, render instead of exporting
 if (typeof window !== 'undefined') {
-	render(appElement, document.getElementById('root'));
+	render(
+		appElement,
+		document.getElementById('root')
+	);
 }
 
 export default appElement;
