+++ pattern-activity-panels/src/index.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import {render} from 'react-dom';
+
 import App from './App';
 
 const appElement = (<App />);
