+++ pattern-expandablelist-object/package.json
@@ -1,40 +1,41 @@
 {
-	"name": "pattern-expandablelist-object",
-	"version": "1.0.0",
-	"description": "A sample demonstrating ExpandableList with data in Object format in an Enact Moonstone application.",
-	"author": "",
-	"main": "src/index.js",
-	"scripts": {
-		"serve": "enact serve",
-		"pack": "enact pack",
-		"pack-p": "enact pack -p",
-		"watch": "enact pack --watch",
-		"clean": "enact clean",
-		"lint": "enact lint --strict .",
-		"test": "enact test start --single-run --browsers PhantomJS",
-		"test-json": "enact test start --single-run --browsers PhantomJS --reporters json",
-		"test-watch": "enact test start --browsers PhantomJS"
-	},
-	"license": "Apache-2.0",
-	"private": true,
-	"repository": "",
-	"enact": {
-		"isomorphic": true,
-		"ri": {
-			"baseSize": 24
-		}
-	},
-	"eslintConfig": {
-		"extends": "enact/strict"
-	},
-	"dependencies": {
-		"@enact/core": "^1.2.0",
-		"@enact/ui": "^1.2.0",
-		"@enact/moonstone": "^1.2.0",
-		"@enact/spotlight": "^1.2.0",
-		"@enact/i18n": "^1.2.0",
-		"@enact/webos": "^1.2.0",
-		"react": "^15.5.4",
-		"react-dom": "^15.5.4"
-	}
+  "name": "pattern-expandablelist-object",
+  "version": "1.0.0",
+  "description": "A sample demonstrating ExpandableList with data in Object format in an Enact Moonstone application.",
+  "author": "",
+  "main": "src/index.js",
+  "scripts": {
+    "serve": "enact serve",
+    "pack": "enact pack",
+    "pack-p": "enact pack -p",
+    "watch": "enact pack --watch",
+    "clean": "enact clean",
+    "lint": "enact lint --strict .",
+    "test": "enact test start --single-run --browsers PhantomJS",
+    "test-json": "enact test start --single-run --browsers PhantomJS --reporters json",
+    "test-watch": "enact test start --browsers PhantomJS"
+  },
+  "license": "Apache-2.0",
+  "private": true,
+  "repository": "",
+  "enact": {
+    "isomorphic": true,
+    "ri": {
+      "baseSize": 24
+    }
+  },
+  "eslintConfig": {
+    "extends": "enact/strict"
+  },
+  "dependencies": {
+    "@enact/core": "next",
+    "@enact/i18n": "next",
+    "@enact/moonstone": "next",
+    "@enact/spotlight": "next",
+    "@enact/ui": "next",
+    "@enact/webos": "next",
+    "prop-types": "^15.6.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0"
+  }
 }
