+++ pattern-locale-switching/src/main.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+import App from './App';
+import configureStore from './store';
+import {Provider} from 'react-redux';
+
+// set default launch path
+const store = configureStore();
+
+let appElement = () => (
+	<Provider store={store}>
+		<App />
+	</Provider>
+);
+
+export default appElement;
