+++ pattern-video-player/package.json
@@ -29,14 +29,14 @@
     "extends": "enact/strict"
   },
   "dependencies": {
-    "@enact/core": "^1.3.1",
-    "@enact/i18n": "^1.3.1",
-    "@enact/moonstone": "^1.3.1",
-    "@enact/spotlight": "^1.3.1",
-    "@enact/ui": "^1.3.1",
-    "@enact/webos": "^1.3.1",
-    "prop-types": "^15.5.10",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4"
+    "@enact/core": "next",
+    "@enact/i18n": "next",
+    "@enact/moonstone": "next",
+    "@enact/spotlight": "next",
+    "@enact/ui": "next",
+    "@enact/webos": "next",
+    "prop-types": "^15.6.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0"
   }
 }
