+++ pattern-virtualgridlist-api/src/components/ImageItem/ImageItem.js
@@ -20,7 +20,6 @@ const ImageItem = kind({
 	},
 
 	render: ({caption, selected, selectImageItem, selectionOverlayShowing, source, subCaption, ...rest}) => {
-		delete rest.data;
 		delete rest.index;
 
 		return (
