+++ pattern-virtualgridlist-api/src/index.js
@@ -1,20 +1,16 @@
 import React from 'react';
 import {render} from 'react-dom';
-import {Provider} from 'react-redux';
-import App from './App';
-import configureStore from './store';
 
-const store = configureStore();
+import App from './main';
 
-const appElement = (
-	<Provider store={store}>
-		<App />
-	</Provider>
-);
+let appElement = <App />;
 
-// In a browser environment, render the app to the document.
+// In a browser environment, render instead of exporting
 if (typeof window !== 'undefined') {
-	render(appElement, document.getElementById('root'));
+	render(
+		appElement,
+		document.getElementById('root')
+	);
 }
 
 export default appElement;
