+++ pattern-virtuallist-preserving-focus/package.json
@@ -1,44 +1,44 @@
 {
-	"name": "pattern-virtuallist-preserving-focus",
-	"version": "1.0.0",
-	"description": "An example of preserving focus in VirtualList for an Enact Moonstone application.",
-	"author": "",
-	"main": "src/index.js",
-	"scripts": {
-		"serve": "enact serve",
-		"pack": "enact pack",
-		"pack-p": "enact pack -p",
-		"watch": "enact pack --watch",
-		"clean": "enact clean",
-		"lint": "enact lint --strict .",
-		"license": "enact license",
-		"test": "enact test start --single-run --browsers PhantomJS",
-		"test-json": "enact test start --single-run --browsers PhantomJS --reporters json",
-		"test-watch": "enact test start --browsers PhantomJS"
-	},
-	"license": "Apache-2.0",
-	"private": true,
-	"repository": "",
-	"enact": {
-		"isomorphic": true,
-		"ri": {
-			"baseSize": 24
-		}
-	},
-	"eslintConfig": {
-		"extends": "enact/strict"
-	},
-	"dependencies": {
-		"@enact/core": "^1.4.1",
-		"@enact/ui": "^1.4.1",
-		"@enact/moonstone": "^1.4.1",
-		"@enact/spotlight": "^1.4.1",
-		"@enact/i18n": "^1.4.1",
-		"@enact/webos": "^1.4.1",
-		"prop-types": "^15.5.8",
-		"react": "^15.5.4",
-		"react-dom": "^15.5.4",
-		"react-redux": "^5.0.3",
-		"redux": "^3.6.0"
-	}
+  "name": "pattern-virtuallist-preserving-focus",
+  "version": "1.0.0",
+  "description": "An example of preserving focus in VirtualList for an Enact Moonstone application.",
+  "author": "",
+  "main": "src/index.js",
+  "scripts": {
+    "serve": "enact serve",
+    "pack": "enact pack",
+    "pack-p": "enact pack -p",
+    "watch": "enact pack --watch",
+    "clean": "enact clean",
+    "lint": "enact lint --strict .",
+    "license": "enact license",
+    "test": "enact test start --single-run --browsers PhantomJS",
+    "test-json": "enact test start --single-run --browsers PhantomJS --reporters json",
+    "test-watch": "enact test start --browsers PhantomJS"
+  },
+  "license": "Apache-2.0",
+  "private": true,
+  "repository": "",
+  "enact": {
+    "isomorphic": true,
+    "ri": {
+      "baseSize": 24
+    }
+  },
+  "eslintConfig": {
+    "extends": "enact/strict"
+  },
+  "dependencies": {
+    "@enact/core": "next",
+    "@enact/i18n": "next",
+    "@enact/moonstone": "next",
+    "@enact/spotlight": "next",
+    "@enact/ui": "next",
+    "@enact/webos": "next",
+    "prop-types": "^15.6.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "react-redux": "^5.0.3",
+    "redux": "^3.6.0"
+  }
 }
